Add missing tests
